+++ data/db/migrations/containers/migration_2025-08-28_0028_process_instance_log.php
@@ -1,24 +1,22 @@
 <?php
 
-namespace App\Data\Db\Migrations;
+namespace App\Data\Db\Migrations\Containers;
 
-use App\Core\DB\ABaseMigration;
+use App\Core\DB\AContainerBaseMigration;
 use App\Core\DB\Helpers\TableSchema;
 use App\Core\DB\Helpers\TableSeeding;
 
-class migration_2025_08_28_0027_application_database_logger extends ABaseMigration {
+class migration_2025_08_28_0028_process_instance_log extends AContainerBaseMigration {
     public function up(): TableSchema {
         $schema = $this->getTableSchema();
 
-        $schema->create('application_log')
+        $schema->create('process_instance_log')
             ->primaryKey('logId')
+            ->varchar('instanceId')
             ->text('message')
-            ->text('stackTrace', true)
-            ->varchar('userId', 256, true)
-            ->varchar('type')
-            ->varchar('method')
+            ->varchar('userId')
+            ->varchar('tsDateCreated')
             ->datetimeAuto('dateCreated')
-            ->integer('tsCreated', 64)
         ;
 
         return $schema;