+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Here is the list of changes in versions.
 - Added support for automatic user deleting
     - Users are deleted after the retention period has ended
 - Added logging to database
+- Added process transaction log
 - Reworked process reports
 - Reworked container permanent flash messages
 - Fixed the users grid in Superadministration settings